validate: Add hint about suggested related categories
Now that related categories are not required (in general, see below for
an exception), we don't need to error out when they're missing. Still,
it's nice to suggest them with a hint.

Note that there are still cases where another category is required
(AudioVideo for Audio, for instance).

Part of https://bugs.freedesktop.org/show_bug.cgi?id=35844 (comment 6
and later)