desktop-file-validate: Set locale so that output encoding is correct
The typical case is that UTF-8 values from a .desktop file are displayed
on a UTF-8 terminal. If we don’t set the locale from the environment,
though, libc assumes that the terminal only supports ASCII, and
substitutes all non-ASCII characters with `?`. This makes (for example)
Cyrillic characters in warning messages useless.

Signed-off-by: Philip Withnall <withnall@endlessm.com>