validate: "Exec" key is required when "DBusActivatable" not set
Before this change, the "Exec" key requirement in the main desktop entry
group was not checked, while an error was always thrown when it was missing
from an action.

After this change, an error will be thrown only if "DBusActivatable" is not
set to true, otherwise a warning will be displayed, following the
specification.

Closes: https://gitlab.freedesktop.org/xdg/desktop-file-utils/-/issues/6